bot/requirements.txt
aiogram
httpx
python-dotenv
pyrogram
